+++ src/actiondoctor/rules/reliability/__init__.py
@@ -1,0 +1,5 @@
+"""Reliability rules."""
+
+from actiondoctor.rules.reliability.rel001_missing_jobs import MissingJobsRule
+
+__all__ = ["MissingJobsRule"]